\name{loggingCommons-package}
\alias{loggingCommons-package}
\alias{loggingCommons}
\docType{package}
\title{
    Logging framework
}
\description{
    Logging framework to support LOG_LEVEL and LOG_DIRECTORY
}
\details{
\tabular{ll}{
Package: \tab loggingCommons\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-02\cr
License: \tab BSD\cr
}
setup logging framework
}
\author{
Irfan Azeez

Maintainer: Irfan Azeez <irfan@itee.uq.edu.au>
}
\references{
}
\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
